requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
snowflake-connector-python
apscheduler
python-dotenv
